Simplify how we deal with payments when returning a sale.
Whenever a sale with multiple installments is being returned, do not cancel
the upcoming installments.
Instead of cancelling a payment, we are creating a outpayment with the
returned value.

We are also removing the last possible step on salereturnwizard.
Also, making some changes on the now the last step of the wizard.
  - Removing total_amount_abs from the ui, so we can avoid some confusion on
    what that information means.
  - Removing the credit_checkbutton
  - Adding radio button to control rather credit will be generated or refund.
  - Adding a label to show when changing the radio button

Update test.

Fixes: https://gitlab.com/stoqtech/stoq/issues/35
Change-Id: I28f79e2948e8713bcf187c3297e266e865daa973
(cherry picked from commit c948985773882b99a35a72319a4d84dacfce4ba2)