Increase pillow dependency to 3.1.0
Previously, when installing pillow, its minimum version should be 1.1.7 and,
the maximum, 4.3.0, otherwise stoq had problems to run. To correct it, found
which attribute on the current version of pillow is the substitute to
deprecated VERSION but exists in both the current version and in the
3.1.0(which is the one used by xenial), changed it on the dependencies.py code
along with pillow version. Also changed requirements.txt and debian/control to
match the minimum version.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/383
Change-Id: I3a138643589e2a3f708b9979c22b400dd6ae1949