Add KPS methods to stoq domain
With this patch, a new method called get_kitchen_items is introduced to
the Sale class. This method checks for all sale items and returns a list
with those whose sellable requires kitchen production.

To do this, a new getter was added to the Sellable class,which returns
whether a sellable requires kitchen production.

Fixes: https://gitlab.com/stoqtech/private/bdil/issues/639

Change-Id: I3f9a21e2d0cb5d713bcd45174367fbe182096216