Include cash fund on money till entry
With this patch, a new parameter was included in stoq in order
to sinalize to our application that the till operations with
money expects the cash fund to be included when opening or closing
the till.

When closing the till, if this new param is true, we will add to
the system_value the initial cash amount for the money method.

Fixes: https://gitlab.com/stoqtech/private/stoq/issues/31
Change-Id: I3706b2baab849a9b4e1416cc1dc05a70c5da4b9f