Fix recursion stack errors
When both configurations are True there are cases
of products that are components of different parents
with different costs that can lead to a recursion loop.

We have seen this onii. This patch adds an extra stopping
condition in one of the 2 recursive functions related to
this cost update chain.

Fix: #59